Last Working
Last working commit for a minute, going to rework the way the main api works to bring it to be more like the browser. Also goingto chnage the way a document is opened. Instead of creating the document around a file, it will create a dom tree and populate it via native commands like createelement and appendchild.