Added state
Added a new state object that gets passed to ComputeStyle. The purpose is to remove the computed state for the rendering process away from the nodes themselves and to pass it to a seperate object. The state is everything for the renderer adapters to render the page. Now ComputeStyle is a true style computer and will give the exact state of the document object as the element properties are.